refactor(agents): dedupe agent-run wiring into shared buildAgentRun + engine singleton (#1777)
server.js's boot-time resume path and routes/agents/runs.js's manual-trigger
and resume-from-terminated handlers each independently resolved a profile's
workflow (embedded vs external ref), wired the wall-time budget/default
model, and applied per-node model/review overrides. Any change to that
wiring had to be made in three places or resumed runs would silently
diverge from freshly-started ones.

Extract the shared resolution into server/agents/runtime/buildAgentRun.js
(also home to applyNodeModels/applyReviewSettings, re-exported from
routes/agents/runs.js for the existing tests) and consolidate the two
separate `new WorkflowEngine({ defaultTimeout: 30 * 60 * 1000 })`
constructions into one singleton in
server/services/workflow/agentEngineSingleton.js, mirroring the existing
getStateManager() pattern.

No behavior change: verified live against the claude-style-agent profile
that manual trigger, resume-from-terminated, and the boot resume path all
resolve the same workflow and durable config (_agentModelConfig,
_agentReviewConfig) as before.
