+++ server/services/workflow/index.js
@@ -35,3 +35,4 @@ export {
   resetExecutionRegistry
 } from './ExecutionRegistry.js';
 export { WorkflowLLMHelper, default as workflowLLMHelper } from './WorkflowLLMHelper.js';
+export { getAgentWorkflowEngine, resetAgentWorkflowEngine } from './agentEngineSingleton.js';
